% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{alphabetize_combos}
\alias{alphabetize_combos}
\title{alphabetize gene combinations}
\usage{
alphabetize_combos(df)
}
\arguments{
\item{df}{any df with columns gene1 and gene2}
}
\description{
alphabetize gene combinations
}
