% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{average_gene_scores}
\alias{average_gene_scores}
\title{Summarize gene combinations by calculating the average lfc for a combo and the
average lfc for each of gene on its own in a combo}
\usage{
average_gene_scores(lfc_df, n_guides = F)
}
\arguments{
\item{lfc_df}{data frame with the columns 'gene1', 'gene2', 'guide1',
'guide2', 'control1', 'control2', ...}

\item{n_guides}{how many guides are in each construct (for grouping controls)}
}
\description{
Summarize gene combinations by calculating the average lfc for a combo and the
average lfc for each of gene on its own in a combo
}
