% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_residuals.R
\name{calculate_anchor_residuals}
\alias{calculate_anchor_residuals}
\title{Calculate genetic interaction scores by treating each guide as an anchor}
\usage{
calculate_anchor_residuals(lfc_df, n_guides = FALSE, fit_controls = FALSE)
}
\arguments{
\item{lfc_df}{data frame with the columns 'gene1', 'gene2', 'guide1',
'guide2', 'control1', 'control2', ...}

\item{n_guides}{guides per gene if FALSE, then controls are not grouped into
pseudogenes}

\item{fit_controls}{if TRUE only use controls to fit linear model}
}
\description{
Calculate genetic interaction scores by treating each guide as an anchor
}
