% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_lfcs.R
\name{calculate_base_lfcs}
\alias{calculate_base_lfcs}
\title{Take the median lfc for each guide paired with controls to establish a
base lfc}
\usage{
calculate_base_lfcs(lfcs)
}
\arguments{
\item{lfcs}{tidy lfcs}
}
\description{
Take the median lfc for each guide paired with controls to establish a
base lfc
}
