% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{check_input}
\alias{check_input}
\title{Checks input for correct column names}
\usage{
check_input(input_df)
}
\arguments{
\item{input_df}{input data}
}
\description{
Checks input for correct column names
}
