% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{cluster_order}
\alias{cluster_order}
\title{return order tibble by hierarchical clustering}
\usage{
cluster_order(scores_df, score)
}
\arguments{
\item{scores_df}{tibble with scores and objects to be clustered
in tidy format.}

\item{score}{score to cluster on}
}
\description{
return order tibble by hierarchical clustering
}
