% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_residuals.R
\name{get_fdrs}
\alias{get_fdrs}
\title{Calculate an empircal fdr combo z-scores}
\usage{
get_fdrs(z_score_df)
}
\arguments{
\item{z_score_df}{combo level z-scores}
}
\description{
Calculate an empircal fdr combo z-scores
}
