% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_residuals.R
\name{get_guide_residuals}
\alias{get_guide_residuals}
\title{Calculate the residual for each guide}
\usage{
get_guide_residuals(rev_base_lfc)
}
\arguments{
\item{dual_base_lfc}{df with every guide pair (guide2) for each guide (guide1)}
}
\description{
Calculate the residual for each guide
}
