% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_residuals.R
\name{get_one_guide_z}
\alias{get_one_guide_z}
\title{Calculate the z-score of the residual of each guide pair}
\usage{
get_one_guide_z(residuals)
}
\arguments{
\item{residuals}{guide level residuals}
}
\description{
Calculate the z-score of the residual of each guide pair
}
