% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_residuals.R
\name{get_two_gene_z}
\alias{get_two_gene_z}
\title{Calculate the gene level z-score from the guide level z-scores of residuals}
\usage{
get_two_gene_z(one_gene_z)
}
\arguments{
\item{one_gene_z}{df from get_one_guide_z}
}
\description{
Calculate the gene level z-score from the guide level z-scores of residuals
}
