% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{group_controls}
\alias{group_controls}
\title{Groups together controls into pseudogenes}
\usage{
group_controls(input_df, guides_per_gene)
}
\arguments{
\item{input_df}{data frame with columns checked by check_input}

\item{n_guides}{number of guides/gene}
}
\description{
Groups together controls into pseudogenes
}
