% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_lfcs.R
\name{join_base_lfcs}
\alias{join_base_lfcs}
\title{Join base lfcs for each guide in a construct}
\usage{
join_base_lfcs(lfcs, base_lfcs)
}
\arguments{
\item{lfcs}{tidy lfcs}

\item{base_lfcs}{- calculated base lfcs}
}
\description{
Join base lfcs for each guide in a construct
}
