% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_combo_residuals}
\alias{plot_combo_residuals}
\title{Visualize the residual plot for all of the guides for a pair of genes}
\usage{
plot_combo_residuals(geneA, geneB, lfcs, n_guides = F)
}
\arguments{
\item{geneA}{first gene to plot}

\item{geneB}{second gene to plot}

\item{lfcs}{log-fold changes}
}
\description{
Visualize the residual plot for all of the guides for a pair of genes
}
