% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_hit_network}
\alias{plot_hit_network}
\title{create a network of top hits with genes as nodes and scores as edges}
\usage{
plot_hit_network(combo_scores, score, cutoff)
}
\arguments{
\item{combo_scores}{scores to plot}

\item{score}{name of the score column}

\item{cutoff}{absolute score cutoff at which to draw edges}
}
\description{
create a network of top hits with genes as nodes and scores as edges
}
