% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_score_heatmap}
\alias{plot_score_heatmap}
\title{plot heatmap of combo scores}
\usage{
plot_score_heatmap(combo_scores, score)
}
\arguments{
\item{combo_scores}{scores to plot}

\item{score}{name of the score column}
}
\description{
plot heatmap of combo scores
}
