% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{reverse_guides}
\alias{reverse_guides}
\title{reverse all guide scores, so we can group by the first gene and summarize}
\usage{
reverse_guides(base_lfcs)
}
\arguments{
\item{base_lfcs}{- tidy lfcs with gene1, gene2, guide1, guide2, base_lfc1,
base_lfc2, control1, control2}
}
\description{
reverse all guide scores, so we can group by the first gene and summarize
}
